[ADDED] New common module for data model used across different modules

+++ app/src/main/java/dev/hossain/weatheralert/db/Alert.kt
@@ -5,7 +5,7 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import dev.hossain.weatheralert.data.WeatherAlertCategory
+import dev.hossain.weatheralert.datamodel.WeatherAlertCategory
 
 /**
  * Represents a user's alert for a specific city.
@@ -28,12 +28,12 @@ import dev.hossain.weatheralert.data.WeatherAlertCategory
     indices = [Index(value = ["city_id"])],
 )
 data class Alert(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     /**
      * Foreign key to City table.
      * @see [City.id]
      */
-    @ColumnInfo(name = "city_id") val cityId: Int,
+    @ColumnInfo(name = "city_id") val cityId: Long,
     @ColumnInfo(name = "alert_category") val alertCategory: WeatherAlertCategory,
     val threshold: Float,
     val notes: String = "",
